fix: adjust correct code structure
