+++ main.py
@@ -8,10 +8,11 @@ from flask_cors import CORS
 
 from logging_module.context import correlation_id
 from logging_module.logger import logger
-from logic import Key_data
-from logic.Key_data import KeyData
+
 from logic.deleter import Deleter
-from logic.migration import Migration
+from logic.migration_mp import MigrationMp
+from logic.migration_ori import MigrationOri
+from logic.migration_sy import MigrationSy
 
 load_dotenv()
 app = Flask(__name__)
@@ -22,7 +23,7 @@ CORS(app)
 def get_migration():
     correlation_id.set(uuid.uuid4())
     logger.info("Info - start_migration service has benn invoked")
-    migration = Migration(os.getenv("TOKEN_PROVIDER"), os.getenv("TOKEN_DESTINATION_WR"))
+    migration = MigrationOri(os.getenv("TOKEN_PROVIDER"), os.getenv("TOKEN_DESTINATION_WR"))
     resp = migration.start_migration()
 
     if resp.default_status != 200:
@@ -34,11 +35,26 @@ def get_migration():
     return resp
 
 
+@app.route('/start_migration_sy/')
+def get_migration_sy():
+    correlation_id.set(uuid.uuid4())
+    logger.info("Info - start_migration_sy service has benn invoked")
+    migration = MigrationSy(os.getenv("TOKEN_PROVIDER"), os.getenv("TOKEN_DESTINATION_WR"))
+    resp = migration.start_migration()
+
+    if resp.default_status != 200:
+        logger.error("An Error has been detected saving origin data at destination, code  error {0}"
+                     .format(resp.status_code))
+        return Response("Error saving origin data at destination", resp.status_code, mimetype="application/json")
+
+    logger.info("Data  migration has been  ended, status  code  {0}".format(resp.status_code))
+    return resp
+
 @app.route('/start_migration_mp/')
 def get_migration_mp():
     correlation_id.set(uuid.uuid4())
     logger.info("Info - start_migration service has benn invoked")
-    migration = Migration(os.getenv("TOKEN_PROVIDER"), os.getenv("TOKEN_DESTINATION_WR"))
+    migration = MigrationMp(os.getenv("TOKEN_PROVIDER"), os.getenv("TOKEN_DESTINATION_WR"))
     resp = migration.start_migration_mp()
 
     if resp.default_status != 200:
